+++ src/st3/r306.cpp
@@ -86,9 +86,9 @@ void R306Init()
         SceExec(0x12, (TaskFunc) r306_checkDoor30bKeyUse, 0, 0, 2, 0);
     }
     if (ItfFlagChk(pG, ITF_R308_THERMO_RIFLE)) {
-        EstSet(0, -1, 0, 0, 1, 1, 1, 0, (u32) zero, zero);
+        EstSet(0, -1, 0, 0, 1, 1, 1, 0, zero, zero);
     } else {
-        EstSet(0, -1, 0, 0, 1, 0, 1, 0, (u32) zero, zero);
+        EstSet(0, -1, 0, 0, 1, 0, 1, 0, zero, zero);
     }
     if (ScfFlagChk(pG, SCF_R307_REGENERATER_APPEAR)) {
         r306_work->em[0].setEm(0x30, 6, 0, 1, 1);
